requirements.txt
flask==1.1.2
flask_socketio==5.1.1
openai==0.27.0
pyyaml==5.4.1
bootstrap==5.1.3